+++ ImpactfulSkills/ImpactfulSkills.cs
@@ -21,7 +21,7 @@ namespace ImpactfulSkills
     {
         public const string PluginGUID = "MidnightsFX.ImpactfulSkills";
         public const string PluginName = "ImpactfulSkills";
-        public const string PluginVersion = "0.5.17";
+        public const string PluginVersion = "0.5.18";
 
         public ValConfig cfg;
         // Use this class to add your own localization to the game